Reversing to unsigned int and fixing Betty issues

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -1,5 +1,4 @@
 #include "lists.h"
-
 /**
  * print_listint_safe - prints a linked
  * list, safely
@@ -8,7 +7,7 @@
  */
 size_t print_listint_safe(const listint_t *head)
 {
-	size_t num_of_nodes = 0;
+	unsigned int num_of_nodes = 0;
 	long int diff;
 
 	while (head)
